# swagger.yaml
openapi: 3.0.3
info:
  title: Nutrition Tracker API
  description: >
    RESTful API for nutrition tracking application with user management, 
    food library with macronutrients, daily entries with meal types, and calorie/macro tracking.
    
    
    Data Model: User has many Days, Day has many Entries, Entry references one Food and has a MealType.
    
    
    Authentication: Most endpoints require a JWT token in the Authorization header (Bearer token format).
    
    
    Token Refresh: All protected endpoints return a refreshed token in the X-Refreshed-Token response header.
  version: 2.0.0
  contact:
    name: API Support
    email: de225789@ucf.edu
servers:
  - url: http://localhost:5000/api
    description: Local development server
  - url: https://colorsdigitalocean.xyz/api
    description: Production server
...
paths:
...
  /users/{userId}/days/{dayId}/entries:
    post:
      tags:
        - Entries
      summary: Add an entry
      description: Add a food entry to a day with meal type (via barcode scan or manual input)
      operationId: addEntry
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          example: 1
        - name: dayId
          in: path
          required: true
          schema:
            type: integer
          example: 15
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - foodId
                - amount
                - mealType
              properties:
                foodId:
                  type: integer
                  example: 42
                  description: ID of the food being logged
                amount:
                  type: number
                  format: float
                  minimum: 0
                  example: 2.5
                  description: Quantity of units consumed
                mealType:
                  type: string
                  enum: [Breakfast, Lunch, Dinner, Snack]
                  example: Lunch
                  description: Type of meal this entry belongs to
      responses:
        '201':
          description: Entry added successfully with enriched food details
          headers:
            X-Refreshed-Token:
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Entry added successfully
                  entry:
                    $ref: '#/components/schemas/EnrichedEntry'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Day or Food not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{userId}/days/{dayId}/entries/{entryId}:
    patch:
      tags:
        - Entries
      summary: Update an entry
      description: Update the food, amount, or meal type of an entry
      operationId: updateEntry
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          example: 1
        - name: dayId
          in: path
          required: true
          schema:
            type: integer
          example: 15
        - name: entryId
          in: path
          required: true
          schema:
            type: string
          description: MongoDB ObjectId
          example: "507f1f77bcf86cd799439011"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                foodId:
                  type: integer
                  example: 43
                amount:
                  type: number
                  format: float
                  example: 3.0
                mealType:
                  type: string
                  enum: [Breakfast, Lunch, Dinner, Snack]
                  example: Dinner
      responses:
        '200':
          description: Entry updated successfully with enriched food details
          headers:
            X-Refreshed-Token:
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Entry updated successfully
                  entry:
                    $ref: '#/components/schemas/EnrichedEntry'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Day, Entry, or Food not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Entries
      summary: Delete an entry
      description: Remove a food entry from a day
      operationId: deleteEntry
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          example: 1
        - name: dayId
          in: path
          required: true
          schema:
            type: integer
          example: 15
        - name: entryId
          in: path
          required: true
          schema:
            type: string
          example: "507f1f77bcf86cd799439011"
      responses:
        '200':
          description: Entry deleted successfully
          headers:
            X-Refreshed-Token:
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Entry deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Day or Entry not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /login endpoint
  schemas:
...
    EnrichedEntry:
      type: object
      description: Entry with calculated macros and food details for Flutter frontend
      properties:
        entryId:
          type: string
          example: "507f1f77bcf86cd799439011"
        foodId:
          type: integer
          example: 42
        foodName:
          type: string
          example: Grilled Chicken Breast
        amount:
          type: number
          format: float
          example: 2.5
        mealType:
          type: string
          enum: [Breakfast, Lunch, Dinner, Snack]
          example: Lunch
        calories:
          type: integer
          example: 413
          description: Total calories (caloriesPerUnit * amount), rounded
        protein:
          type: integer
          example: 78
          description: Total protein in grams (proteinPerUnit * amount), rounded
        carbs:
          type: integer
          example: 0
          description: Total carbs in grams (carbsPerUnit * amount), rounded
        fat:
          type: integer
          example: 9
          description: Total fat in grams (fatPerUnit * amount), rounded
        timestamp:
          type: string
          format: date-time
          example: "2025-10-30T14:23:00.000Z"
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Validation error message
        success:
          type: boolean
          example: false
  responses:
    UnauthorizedError:
      description: Authentication token is missing or invalid
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: The JWT is no longer valid